requirements.txt
streamlit>=1.32.0
pymupdf>=1.23.0
easyocr>=1.7.0
marker-pdf>=0.2.0
openai>=1.0.0
PyGithub>=2.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0
